requirements.txt
telethon
python-dotenv
python-dotenv
